% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-png-info.R
\name{get_png_info}
\alias{get_png_info}
\title{Get information about a PNG file}
\usage{
get_png_info(src)
}
\arguments{
\item{src}{filename or raw vector containing PNG data}
}
\description{
Get information about a PNG file
}
