% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{normalize_colors}
\alias{normalize_colors}
\title{Normalize named R colors to hex colors of the form \code{#RRGGBB}}
\usage{
normalize_colors(x, alpha = FALSE)
}
\arguments{
\item{x}{vector, matrix or raster which includes named colors like 'white'}

\item{alpha}{logical. Should the alpha value be included}
}
\value{
new data with named R colors replaced by their hex value
}
\description{
Normalize named R colors to hex colors of the form \code{#RRGGBB}
}
\examples{
normalize_colors(matrix(c('red', 'white', '#444444'), 3, 4))
normalize_colors(matrix(c('red', 'white', '#444444'), 3, 4), alpha = TRUE)
}
