% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-ihdr.R
\name{png_info}
\alias{png_info}
\title{Get information about a PNG file}
\usage{
png_info(raw_vec)
}
\arguments{
\item{raw_vec}{raw vector containing encoded PNG data}
}
\description{
Get information about a PNG file
}
