% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write-png.R
\name{raw_spec}
\alias{raw_spec}
\title{Create a specification for how raw bytes should be interpreted when passed
to \code{write_png()}}
\usage{
raw_spec(width, height, depth, bits)
}
\arguments{
\item{width, height}{image dimensions}

\item{depth}{number of colour channels. Integer value in range [1, 4]}

\item{bits}{number of bits for each colour channel. Either 8 or 16.}
}
\value{
named list to pass to the \code{write_png(..., raw_spec = )}
}
\description{
Create a specification for how raw bytes should be interpreted when passed
to \code{write_png()}
}
\examples{
raw_spec(100, 20, 3, 8)
}
