% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-png.R
\name{read_png_raw}
\alias{read_png_raw}
\alias{read_png_nara}
\alias{read_png_raster}
\alias{read_png_rgba}
\alias{read_png_rgb}
\title{Read a PNG from a raw vector in memory}
\usage{
read_png_raw(raw_vec, fmt = spng_format$SPNG_FMT_RGBA8, flags = 0)

read_png_nara(raw_vec, flags = 0)

read_png_raster(raw_vec, flags = 0)

read_png_rgba(raw_vec, flags = 0)

read_png_rgb(raw_vec, flags = 0)
}
\arguments{
\item{raw_vec}{raw vector containing encoded PNG data}

\item{fmt}{what format the decoded bytes should be in. default: \code{SPNG_FMT_RGBA8}.
See \code{?spng_format} for other formats.}

\item{flags}{flags. default: 0. See \code{?spng_decode_flags} for other
options.  Must be an integer.}
}
\value{
vector of raw bytes representing pixel data in the requested format.
        Note that the pixels will be in row-major format
}
\description{
See \url{https://libspng.org/docs/decode/} for more information on which
combinations of formats and flags are permissable.
}
