% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\docType{data}
\name{spng_decode_flags}
\alias{spng_decode_flags}
\title{Flags for decoding}
\format{
An object of class \code{list} of length 2.
}
\usage{
spng_decode_flags
}
\description{
\describe{
  \item{SPNG_DECODE_TRNS}{Apply transparency}
  \item{SPNG_DECODE_GAMMA}{Apply gamma correction}
}
}
\keyword{datasets}
