% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\docType{data}
\name{spng_format}
\alias{spng_format}
\title{Decode PNG as the specified format type}
\format{
An object of class \code{list} of length 8.
}
\usage{
spng_format
}
\description{
A good default is \code{SPNG_FMT_RGBA8}
}
\details{
\describe{
  \item{SPNG_FMT_RGBA8  }{RGBA color.  8-bit color. Convert from any PNG format and bit depth to this format}
  \item{SPNG_FMT_RGBA16 }{RGBA color.  8-bit color. Convert from any PNG format and bit depth to this format}
  \item{SPNG_FMT_RGB8   }{RGB color . 16-bit color. Convert from any PNG format and bit depth to this format}
  \item{SPNG_FMT_GA8    }{Only valid for 1, 2, 4, 8-bit grayscale PNGs}
  \item{SPNG_FMT_GA16   }{Only valid for 16-bit grayscale PNGs}
  \item{SPNG_FMT_G8     }{Only valid for 1, 2, 4, 8-bit grayscale PNGs}
  \item{SPNG_FMT_PNG    }{The PNG’s format in host-endian}
  \item{SPNG_FMT_RAW    }{The PNG’s format in big-endian}
}
}
\keyword{datasets}
