% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-images.R
\docType{data}
\name{test_image}
\alias{test_image}
\title{A named list of test images arranged by storage type: array, raster, 
native_raster}
\format{
An object of class \code{list} of length 9.
}
\usage{
test_image
}
\description{
A named list of test images arranged by storage type: array, raster, 
native_raster
}
\keyword{datasets}
