% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-images.R
\docType{data}
\name{test_image_gray8}
\alias{test_image_gray8}
\title{Test image with grayscale values in a 2D matrix}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 150 rows and 200 columns.
}
\usage{
test_image_gray8
}
\description{
Test image with grayscale values in a 2D matrix
}
\keyword{datasets}
