% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-images.R
\docType{data}
\name{test_image_raster}
\alias{test_image_raster}
\title{Test image with RGBA hex colour values values in a 2D character matrix}
\format{
An object of class \code{raster} with 150 rows and 200 columns.
}
\usage{
test_image_raster
}
\description{
Test image with RGBA hex colour values values in a 2D character matrix
}
\keyword{datasets}
