% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write-png.R
\name{write_png_from_nara}
\alias{write_png_from_nara}
\title{Write PNG from NativeRaster}
\usage{
write_png_from_nara(
  nara,
  file = NULL,
  use_filter = TRUE,
  compression_level = -1L
)
}
\arguments{
\item{nara}{native raster object. I.e.  An integer matrix.}

\item{file}{If NULL then return result as raw vector, otherwise write
to the given filepath.}

\item{use_filter}{Use PNG filtering to help reduce size. Default: TRUE.
If FALSE, then filtering will be disabled which can make 
image writing faster.}

\item{compression_level}{compression level for zlib.  Default: -1 means
to use the default compression level.  Other valid
values are in range [0, 9].  In general, lower compression levels
result in faster compression, but larger image sizes.  For fastest
image writing, set \code{compression_level}
to 0 to completely disable compression.}
}
\description{
Write PNG from NativeRaster
}
