% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-png-info.R
\name{extract_png_info}
\alias{extract_png_info}
\title{Get information about a PNG file}
\usage{
extract_png_info(raw_vec)
}
\arguments{
\item{raw_vec}{raw vector containing encoded PNG data}
}
\description{
Get information about a PNG file
}
