% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-png.R
\name{read_png}
\alias{read_png}
\title{Read a PNG from a raw vector in memory}
\usage{
read_png(src, type = c("nara", "raster", "rgba", "rgb"), flags = 0L)
}
\arguments{
\item{src}{raw vector containing encoded PNG data or file path}

\item{type}{type of object in which to store image data}

\item{flags}{flags. default: 0. See \code{?spng_decode_flags} for other
options.  Must be an integer.}
}
\value{
R image object of the specified type
}
\description{
See \url{https://libspng.org/docs/decode/} for more information on which
combinations of formats and flags are permissable.
}
