% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-png.R
\name{read_png}
\alias{read_png}
\title{Read a PNG from a raw vector in memory}
\usage{
read_png(
  src,
  type = c("array", "raster", "native_raster", "indexed", "raw"),
  rgba = FALSE,
  flags = 1L,
  avoid_transpose = FALSE,
  array_type = c("dbl", "int")
)
}
\arguments{
\item{src}{raw vector containing encoded PNG data or file path}

\item{type}{type of object in which to store image data. Valid types are
'array', 'raster', 'native_raster', 'indexed' and 'raw'.  Note that indexed
image objects can only be loaded from indexed PNGs.}

\item{rgba}{Should the result be forced into RGBA?  Default: FALSE
Example 1: When reading with \code{type = 'native_raster'} the result
will always be RGBA, so this variable is ignored.
Example 2: When reading a grey+alpha image, if \code{rgba = TRUE}, 
then image will be read into an RGBA array with 4 planes.  Otherwise
it will be read in as grey+alpha data in an array with 2 planes.}

\item{flags}{flags. default: 1 (always decode transparency from tRNS chunks). 
See \code{?spng_decode_flags} for other options.  Must be an integer.}

\item{avoid_transpose}{default: FALSE.}

\item{array_type}{'dbl' or 'int'}
}
\value{
R image object of the specified type
}
\description{
See \url{https://libspng.org/docs/decode/} for more information on which
combinations of formats and flags are permissable.
}
